comments, refined d_path size
